move mock app to non-jsx
